% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda.vi-package.R
\docType{package}
\name{fda.vi-package}
\alias{fda.vi}
\alias{fda.vi-package}
\title{fda.vi: Functional Data Analysis using Variational Inference}
\description{
Implements a variational Expectation-Maximization (VEM) algorithm for smoothing one or multiple functional observations via basis function selection. The algorithm estimates all model parameters simultaneously and automatically, while accounting for within-curve correlation. The approach provides a flexible and computationally efficient framework for smoothing correlated functional data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/desouzalab/fda.vi}
  \item Report bugs at \url{https://github.com/desouzalab/fda.vi/issues}
}

}
\author{
\strong{Maintainer}: Camila de Souza \email{camila.souza@uwo.ca}

Authors:
\itemize{
  \item Stephen Kinsey
  \item Ana Carolina da Cruz
  \item Pedro Henrique Toledo Oliveira Sousa
}

}
\keyword{internal}
